создание REST API

+++ src/main/java/ru/homework3/mvc/service/UserService.java
@@ -2,11 +2,11 @@ package ru.homework3.mvc.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import ru.homework3.mvc.controller.ResponseCode;
+import ru.homework3.mvc.dto.UserDto;
 import ru.homework3.mvc.model.User;
 import ru.homework3.mvc.repo.UserRepository;
-import ru.homework3.mvc.utils.validator.Validator;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -14,31 +14,33 @@ import java.util.List;
 public class UserService {
     private final UserRepository userRepository;
 
-    public String getUsers() {
-        String usersStr = "";
+    public List<UserDto> getUsers() {
         List<User> users = userRepository.findAll();
+        List<UserDto> dtos = new ArrayList<>(users.size());
         for (User user: users) {
-            usersStr += user.toString() +
-                    " {<a href='/" +
-                    user.getId() +
-                    "?filter=1'>Вывести задачи</a> | <a href='" +
-                    user.getId() +
-                    "/delete_user'>Удалить</a>}<br/>";
+            dtos.add(UserDto.build(user));
         }
-        return usersStr;
+        return dtos;
     }
 
-    public ResponseCode createUser(User user) {
-        if (Validator.validateName(user.getName())) {
-            userRepository.save(user);
-            return ResponseCode.SUCCESS;
+    public UserDto getUser(Integer idUser) {
+        if (userRepository.existsById(idUser)) {
+            return UserDto.build(userRepository.getReferenceById(idUser));
         } else {
-            return ResponseCode.ERROR_VALIDATE;
+            return null;
         }
     }
 
-    public void deleteUser(Integer idUser) {
-        userRepository.deleteById(idUser);
+    public UserDto createUser(UserDto dto) {
+        return UserDto.build(userRepository.save(dto.toUser()));
+    }
+
+    public String deleteUser(Integer idUser) {
+        if (userRepository.existsById(idUser)) {
+            userRepository.deleteById(idUser);
+            return "Done!";
+        }
+        return "Fail!";
     }
 
     public void deleteUsers() {
